requirements.txt
playwright
requests
python-dotenv
synology-api